Move functionality for determining the order of operations into separate file.
